requirements.txt
beautifulsoup4==4.13.5
markdownify==1.2.0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5